+++ CHANGELOG.md
@@ -4,4 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 
+## [0.2.0] - Unreleased
+
+- Added two draft specifications.
+- Added structural validator.
+- Updated lexicon to 1.0.2.
+
 ## [0.1.0] - Unreleased